+++ mayan/apps/user_management/icons.py
@@ -19,5 +19,4 @@ icon_user_list = Icon(driver_name='fontawesome', symbol='user')
 icon_user_multiple_delete = icon_user_delete
 icon_user_multiple_set_password = Icon(driver_name='fontawesome', symbol='key')
 icon_user_set_options = Icon(driver_name='fontawesome', symbol='cog')
-icon_user_set_password = Icon(driver_name='fontawesome', symbol='key')
 icon_user_setup = Icon(driver_name='fontawesome', symbol='user')