+++ mayan/apps/user_management/tests/mixins.py
@@ -312,22 +312,3 @@ class UserViewTestMixin(object):
             viewname='user_management:user_groups',
             kwargs={'pk': self.test_user.pk}
         )
-
-    def _request_test_user_password_set_view(self, password):
-        return self.post(
-            viewname='user_management:user_set_password',
-            kwargs={'pk': self.test_user.pk},
-            data={
-                'new_password1': password, 'new_password2': password
-            }
-        )
-
-    def _request_test_user_password_set_multiple_view(self, password):
-        return self.post(
-            viewname='user_management:user_multiple_set_password',
-            data={
-                'id_list': self.test_user.pk,
-                'new_password1': password,
-                'new_password2': password
-            }
-        )