+++ HISTORY.rst
@@ -244,6 +244,7 @@
 * Add the preparestatic command.
 * Remove the related attribute of check_access.
 * Remove filter_by_access. Replaced by restrict_queryset.
+* Move the user set password views to the authentication app.
 
 3.1.11 (2019-04-XX)
 ===================